Restore previous window layout
For example, I have buffer/window A displayed in a frame, and buffer/window B is
created (or occupied) by atomic-chrome.

Before:

A -> A|B -> A|A

After:

A -> A|B -> A

I make the new behavior only takes effect when the option
`atomic-chrome-buffer-open-style' is 'split (the default) because I have no
knowledage about the other possible options.